+++ crates/simplexpr/src/eval.rs
@@ -501,14 +501,26 @@ fn call_expr_function(name: &str, args: Vec<DynVal>) -> Result<DynVal, EvalError
 
                 Ok(DynVal::from(match timezone.timestamp_opt(timestamp.as_i64()?, 0) {
                     LocalResult::Single(t) | LocalResult::Ambiguous(t, _) => {
-                        t.format_localized(&format.as_string()?, get_locale()).to_string()
+                        let format = format.as_string()?;
+                        let delayed_format = t.format_localized(&format, get_locale());
+                        let mut buffer = String::new();
+                        if delayed_format.write_to(&mut buffer).is_err() {
+                            return Err(EvalError::ChronoError("Invalid time formatting string: ".to_string() + &format));
+                        }
+                        buffer
                     }
                     LocalResult::None => return Err(EvalError::ChronoError("Invalid UNIX timestamp".to_string())),
                 }))
             }
             [timestamp, format] => Ok(DynVal::from(match Local.timestamp_opt(timestamp.as_i64()?, 0) {
                 LocalResult::Single(t) | LocalResult::Ambiguous(t, _) => {
-                    t.format_localized(&format.as_string()?, get_locale()).to_string()
+                    let format = format.as_string()?;
+                    let delayed_format = t.format_localized(&format, get_locale());
+                    let mut buffer = String::new();
+                    if delayed_format.write_to(&mut buffer).is_err() {
+                        return Err(EvalError::ChronoError("Invalid time formatting string: ".to_string() + &format));
+                    }
+                    buffer
                 }
                 LocalResult::None => return Err(EvalError::ChronoError("Invalid UNIX timestamp".to_string())),
             })),